Move closeSlot into slots.js
