+++ functions/errors.js
@@ -1,0 +1,10 @@
+class HTTPError extends Error {
+  constructor(code, message) {
+    super(message);
+    this.code = code;
+  }
+}
+
+module.exports = {
+  HTTPError: HTTPError,
+};
